requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.2.25
django-cors-headers==4.1.0
djangorestframework==3.15.1
gunicorn==23.0.0
importlib-metadata==6.7.0
mysql-connector-python==8.0.33
packaging==24.0
Pillow==9.5.0
protobuf==3.20.3
psycopg2-binary==2.9.9
python-dotenv==0.21.1
pytz==2024.2
sqlparse==0.4.4
typing_extensions==4.7.1
waitress==2.1.2
zipp==3.15.0
